refactor: rework and simplify the entire repository/package
This is just a start on where I’d like to take this package (now `remark-loader`),
and in hindsight forking `react-markdown-loader` wasn’t necessarily
the best approach as I removed a lot and am preparing to port to remark
for markdown parsing. I also removed some of the extra style-guide specific
and instead allow a `template` property to the front-matter which enables
a variety of customization.