chore: remove react specific code
This is part of what is noted towards the end of #1. Now that MDX is a
stable option for integrating react with remark, the goal of this
package will be to provide a barebones `remark` loader with built-in
image resolution.

BREAKING CHANGE: The `react` option has been removed. Our implementation
was extremely hacky and led to a bunch of pitfalls. Please use MDX as an
alternative.